- Properties must now be allocated with new - some more work on PropertyWithMainThreadDelegate

+++ boden/include/windows-classic/bdn/WindowTextProperty.h
@@ -77,7 +77,7 @@ protected:
 	{
 		// this is called while the mutex is locked.
 		// So there are no race conditions here.
-		int		updateId = _nextUpdateId;
+		int64_t		updateId = _nextUpdateId;
 		_nextUpdateId++;
 
 		String	newValue = _value;
